% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{df2table}
\alias{df2table}
\title{Generate catch table for report}
\usage{
df2table(df, spcs = NULL, year, format = "html")
}
\arguments{
\item{df}{Data frame with
\code{c("Year", "Species", "Prefec", "Month", "Catch_ton")}}

\item{spcs}{Character of species name to extract}

\item{year}{Year to extract in numeric}
}
\description{
Generate catch table for report
}
\examples{
\dontrun{
make_table(load_catch_dummy_kagoshima(),
           spcs = "\u30de\u30a2\u30b8", year = 2019)
make_table(stockdbr::load_catch(), spcs = "\u30de\u30a2\u30b8", year = 2019)
}
}
