% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_graph.R
\name{each_graph}
\alias{each_graph}
\title{Draw a time series catch plot for each region}
\usage{
each_graph(data, regional)
}
\arguments{
\item{data}{Tidy catch data sorted by the business requirements}

\item{Regional}{Name of a Region separated by a boundary_prefec}
}
\value{
ggplot figure
}
\description{
Draw a time series catch plot for each region
}
