% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize.R
\name{finalize_df}
\alias{finalize_df}
\title{Finalize data.frame to be in required format}
\usage{
finalize_df(dat)
}
\arguments{
\item{dat}{Data to be processed}
}
\description{
Finalize data.frame to be in required format
}
