% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmtbl.R
\name{fmtbl}
\alias{fmtbl}
\title{Load and format bl histogram data}
\usage{
fmtbl(path, spcs, nest = FALSE)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}

\item{spcs}{Spcs name in romaji, one of}

\item{nest}{If \code{TRUE}, data will be shown in rectangle format
  whith nested bl datafor quick overview.
\itemize{
  \item maiwashi
  \item maaji
  \item sabarui
  \item masaba
  \item gomasaba
  \item katakuchi
  \item urume
}}
}
\description{
Load and format bl histogram data
}
