% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmtbl.R
\name{fmtbl_fresco}
\alias{fmtbl_fresco}
\title{Format bldata exported from FRESCO database}
\usage{
fmtbl_fresco(path, type, date.start, date.end)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}

\item{type}{Format of data to load either 'taichou' or 'seimitsu'.}

\item{date.start}{The first day of the processed data.}

\item{date.end}{The last day of the the processed data}
}
\description{
Format bldata exported from FRESCO database
}
\examples{
\dontrun{
  fmtbl_fresco("2019Mar_seikai_taichou_katakuchi.csv", type = "taichou",
               date.start = "20180901", date.end = "20190331")
  fmtbl_fresco("2019Mar_seikai_seimitsu_katakuchi.csv", type = "seimitsu"
               date.start = "20180901", date.end = "20190331")
}
}
