% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmtcatch_fukuoka.R
\name{fmtcatch.fukuoka}
\alias{fmtcatch.fukuoka}
\title{Load and format catch data of fukuoka}
\usage{
\method{fmtcatch}{fukuoka}(path, spcs, type)
}
\arguments{
\item{path}{File path to process}

\item{spcs}{Romaji spcs name one of
\itemize{
  \item{"maaji"}
  \item{"maiwashi"}
  \item{"sabarui"}
  \item{"katakuchi"}
  \item{"urume"}
}}

\item{type}{Character value either "maki" or "bouuke", and "others".}
}
\description{
Load and format catch data of fukuoka
}
