% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmtcatch_kumamoto_saga.R
\name{fmtcatch.kumamoto}
\alias{fmtcatch.kumamoto}
\title{Load catch data of kumamoto and tidy it up}
\usage{
\method{fmtcatch}{kumamoto}(path, spcs, type)
}
\arguments{
\item{path}{File path to process}

\item{spcs}{Romaji spcs name one of
\itemize {
  \item{"maaji"}
  \item{"maiwashi"}
  \item{"sabarui"}
  \item{"katakuchi"}
  \item{"urume"}
}}

\item{type}{Character value either "maki" or "bouke".
Data processing algorithm is controled by this parameter.}
}
\description{
Load catch data of kumamoto and tidy it up
}
