% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmtcatch.R
\name{fmtcatch}
\alias{fmtcatch}
\title{Load and format catch data}
\usage{
fmtcatch(path, spcs, nest = FALSE)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}

\item{spcs}{Spcs name in romaji, one of}

\item{nest}{If \code{TRUE}, data will be shown in rectangle format
  whith nested catch data for quick overview.
\itemize{
  \item maiwashi
  \item maaji
  \item sabarui
  \item masaba
  \item gomasaba
  \item katakuchi
  \item urume
}}
}
\description{
Load and format catch data
}
