% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmtcatch_yamaguchi.R
\name{fmtcatch.yamaguchi}
\alias{fmtcatch.yamaguchi}
\title{Load and format catch data of yamaguchi}
\usage{
\method{fmtcatch}{yamaguchi}(path, spcs, type = NULL)
}
\arguments{
\item{path}{File path to process}

\item{spcs}{Romaji spcs name one of
\itemize{
  \item{"maaji"}
  \item{"maiwashi"}
  \item{"sabarui"}
  \item{"katakuchi"}
  \item{"urume"}
}}

\item{type}{Character value either "sukui" or "bouuke".}
}
\description{
Load and format catch data of yamaguchi
}
