% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmtbl.R
\name{format_bl}
\alias{format_bl}
\title{Load and format bl histogram data}
\usage{
format_bl(path, prefec, spcs, nest = FALSE)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}

\item{prefec}{Name of prefecture as string, oneof
\itemize{
  \item "yamaguchi"
  \item "fukuoka"
  \item "saga"
  \item "nagasaki"
  \item "kumamoto"
  \item "kagoshima"
}}

\item{spcs}{Spcs name as string, one of
\itemize{
  \item maiwashi
  \item maaji
  \item sabarui
  \item masaba
  \item gomasaba
  \item katakuchi
  \item urume
}}

\item{nest}{If \code{TRUE}, data will be shown in rectangle format
whith nested bl datafor quick overview.}
}
\description{
Load and format bl histogram data
}
