% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmtcatch.R
\name{format_catch}
\alias{format_catch}
\title{Load and format catch data}
\usage{
format_catch(path, prefec, spcs, type = NULL)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}

\item{prefec}{Prefecture name as string, one of
\itemize{
  \item "yamaguchi"
  \item "fukuoka"
  \item "saga"
  \item "nagasaki"
  \item "kumamoto"
  \item "kagoshima"
}}

\item{spcs}{Spcs name in romaji, one of
\itemize{
  \item "maiwashi"
  \item "maaji"
  \item "sabarui"
  \item "masaba"
  \item "gomasaba"
  \item "katakuchi"
  \item "urume"
}}

\item{type}{Character value either "maki" or "bouuke" to control
data processing algorithm for kumamoto data.}
}
\description{
Load and format catch data
}
