% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Denial of \code{\%in\%}}
\usage{
x \%not_in\% y
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.
    \link[base]{Long vectors} are supported.}

\item{y}{vector or NULL: the values to be matched against}
}
\description{
Denial of \code{\%in\%}
}
