% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gyokaikyorize.R
\name{gyokaikyorize}
\alias{gyokaikyorize}
\title{Format and wrangle catch data to fit business requirements}
\usage{
gyokaikyorize(df, species)
}
\arguments{
\item{df}{a data frame to be processed}

\item{species}{species name. Only supported with 'maaji' for now.}
}
\value{
formatted data.frame
}
\description{
Format and wrangle catch data to fit business requirements
}
\examples{
\dontrun{
gyokaikyorize(df = data.frame())
}
}
