% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{make_table}
\alias{make_table}
\title{Generate catch table for report}
\usage{
make_table(df, spcs, year)
}
\arguments{
\item{df}{Data frame with
\code{c("Year", "Species", "Prefec", "Month", "Catch_ton")}}

\item{spcs}{Character of species name to extract}

\item{year}{Year to extract in numeric}
}
\description{
Generate catch table for report
}
\examples{
\dontrun{
make_table(load_catch_dummy_kagoshima(), spcs = "マアジ", year = 2019)
make_table(stockdbr::load_catch(), spcs = "マアジ", year = 2019)
}
}
