% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmtcatch_nagasaki.R
\name{ngs_fmt_sheet}
\alias{ngs_fmt_sheet}
\title{Load catch data from single Excel sheet and tidy it up}
\usage{
ngs_fmt_sheet(sheet, path, regex)
}
\arguments{
\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an
integer (the position of the sheet). Ignored if the sheet is specified via
\code{range}. If neither argument specifies the sheet, defaults to the first
sheet.}

\item{path}{Path to the xls/xlsx file.}

\item{regex}{Regular expression to match the name of the target species.}
}
\description{
Load catch data from single Excel sheet and tidy it up
}
