% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmtcatch_nagasaki.R
\name{ngs_get_colvalue}
\alias{ngs_get_colvalue}
\title{Get values placed separately in column direction}
\usage{
ngs_get_colvalue(regex, df, offset.x = 0, offset.y = 0,
  xtract.digit = FALSE)
}
\arguments{
\item{regex}{Regular expression to match the name of the target species.}

\item{df}{Loaded Excel sheet}

\item{offset.x}{X-offset of the target value
from the position of the month cell}

\item{offset.y}{Y-offest of the target value
from the position of the month cell}

\item{xtract.digit}{If \code{TRUE}, all characters were deleted so that
the return value composed only of numerics.}
}
\description{
This function return values placed separately on the specific fish row
  in column direction.
}
