% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmtcatch_nagasaki.R
\name{ngs_get_monthcol}
\alias{ngs_get_monthcol}
\title{Detect month column coping with variation in row position of the month cell}
\usage{
ngs_get_monthcol(row, df)
}
\arguments{
\item{row}{Assumed position of the month cell}

\item{df}{Loaded Excel sheet}
}
\value{
Row-column set of month cells in a given \code{row}
}
\description{
Detect month column coping with variation in row position of the month cell
}
