% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmtcatch_nagasaki.R
\name{ngs_get_port}
\alias{ngs_get_port}
\title{Get port name from string}
\usage{
ngs_get_port(str)
}
\arguments{
\item{str}{Japanes string which contains port name.}
}
\description{
This function returns port name from remarks written in Excel sheet. Port
name with variants are extracted using regular expression and then
converted to four port names written in ascii characters.
}
