% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmtcatch_nagasaki.R
\name{ngs_locate_spcsrow}
\alias{ngs_locate_spcsrow}
\title{Locate row position of the cell which contains species name}
\usage{
ngs_locate_spcsrow(regex, colstr)
}
\arguments{
\item{regex}{Regular expression to match the name of the target species.}

\item{colstr}{String vector made from a Excel worksheet column.}
}
\value{
All the row positions where the species name start.
}
\description{
This function locates the row position of the cell which contains
  specific species name.
}
