% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmtcatch_nagasaki.R
\name{ngs_make_yrvec}
\alias{ngs_make_yrvec}
\title{Make year vectors considering year change}
\usage{
ngs_make_yrvec(sheet, month)
}
\arguments{
\item{sheet}{Sheet name of Nagasaki catch data (fmt: \%Y.\%m-\%Y.\%m).}

\item{month}{Month vector}
}
\description{
This function makes year vector from Excel sheet name (\%Y.\%m-\%Y.\%m) and
  month vector. If the month vector contains year change (Dec-Jan),
  year vector returned is composed from two years.
}
\examples{
\dontrun{
  ngs_make_yrvec(sheet = "2018.09-2019.03",
                 month = c(9, 10, 11, 12, 1, 2, 3))
}
}
