% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_catch_at_age_legacy.R
\name{parse.catch_at_age_maaji}
\alias{parse.catch_at_age_maaji}
\title{Parse maaji object into data frame}
\usage{
\method{parse}{catch_at_age_maaji}(catch_at_age_maaji)
}
\arguments{
\item{catch_at_age_maaji}{Object instanciated by \code{legacy()}}
}
\description{
Parse maaji object into data frame
}
