% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_legacy.R
\name{parse.catch_maaji}
\alias{parse.catch_maaji}
\title{Parse maaji object into data frame}
\usage{
\method{parse}{catch_maaji}(catch_maaji)
}
\arguments{
\item{catch_maaji}{Object instanciated by \code{legacy()}}
}
\description{
Parse maaji object into data frame
}
