% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_legacy.R
\name{parse.maaji}
\alias{parse.maaji}
\title{Parse maaji object into data frame}
\usage{
\method{parse}{maaji}(maaji)
}
\arguments{
\item{maaji}{Object instanciated by \code{legacy()}}
}
\description{
Parse maaji object into data frame
}
