% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_legacy.R
\name{parse_catch_legacy}
\alias{parse_catch_legacy}
\title{Parse legacy Excel data into data frame}
\usage{
parse_catch_legacy(legacy)
}
\arguments{
\item{legacy}{List which contains
\itemize{
\item{fname}
\item{spcs}
\item{year}
}}
}
\description{
Parse legacy Excel data into data frame
}
