% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_legacy.R
\name{parse_legacy.maiwashi_maaji}
\alias{parse_legacy.maiwashi_maaji}
\title{Parse pseudo database of gyokaikyo project into data frame}
\usage{
parse_legacy.maiwashi_maaji(fname, year, species)
}
\arguments{
\item{fname}{Name of legacy catch database file of maiwashi and maaji}
}
\description{
Note that this function is for maiwashi and maaji
}
