% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch_at_age.R
\name{plot_catch_at_age}
\alias{plot_catch_at_age}
\title{Plot catch at age data}
\usage{
plot_catch_at_age(df, type)
}
\arguments{
\item{df}{Data frame}

\item{type}{"Nourin_adjusted" or "Major_port"}
}
\description{
Plot catch at age data
}
