% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spcsname.R
\name{recode_spcsname}
\alias{recode_spcsname}
\title{Change species name in a specific prefecture}
\usage{
recode_spcsname(dat, prefec, original, newname)
}
\arguments{
\item{dat}{Tidy tibble data from load_catch}

\item{prefec}{Prefecture that contains species name to be changed}

\item{original}{Species name to be changed}

\item{newname}{New species name}
}
\value{
tidy data with new species name
}
\description{
Change species name in a specific prefecture
}
