% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_graph.R
\name{regional_graph}
\alias{regional_graph}
\title{Draw a time vs catch plot separated into two regions}
\usage{
regional_graph(data, species_regex, boundary_prefec, unit = "ton")
}
\arguments{
\item{data}{Tidy catch data sorted by the business requirements}

\item{species_regex}{Species regex to extract target species}

\item{boundary_prefec}{Eastmost prefec of the Western region}

\item{unit}{unit of y-axis chosen from c("ton", "kiloton", "manton")}
}
\value{
ggplot figure
}
\description{
Draw a time vs catch plot separated into two regions
}
