% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seq_log}
\alias{seq_log}
\title{Generate logarithmic sequence}
\usage{
seq_log(from, to = from, by = 1)
}
\arguments{
\item{from}{the starting and (maximal) end values of the
    sequence.  Of length \code{1} unless just \code{from} is supplied as
    an unnamed argument.}

\item{to}{the starting and (maximal) end values of the
    sequence.  Of length \code{1} unless just \code{from} is supplied as
    an unnamed argument.}

\item{by}{number: increment of the sequence.}
}
\description{
Generate logarithmic sequence
}
