% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_region_by_two}
\alias{set_region_by_two}
\title{Separate Prefecs into Eastern and Western regions}
\usage{
set_region_by_two(prefec, boundary_prefec)
}
\arguments{
\item{prefec}{Prefecture name in Japanese from the Pacific block}

\item{boundary_prefec}{Eastmost prefecture of Western Region}
}
\value{
Regionalized prefectures based on the boundary_prefec
}
\description{
Separate Prefecs into Eastern and Western regions
}
