% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery.R
\name{standardize_fishery}
\alias{standardize_fishery}
\title{Standardize duplicated- or molformed- fishery name}
\usage{
standardize_fishery(fishery)
}
\arguments{
\item{fishery}{Column \code{Fishery} of catch data}
}
\description{
Standardize duplicated- or molformed- fishery name
}
