% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spcsname.R
\name{standardize_spcsname}
\alias{standardize_spcsname}
\title{Standardize species name in a database}
\usage{
standardize_spcsname(dat)
}
\arguments{
\item{dat}{Tidy data from load_catch}
}
\value{
tidy data with new species name
}
\description{
Standardize species name in a database
}
