% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmtcatch_nagasaki.R
\name{fmtcatch.nagasaki}
\alias{fmtcatch.nagasaki}
\title{Make catch data reading multiple Excel worksheets}
\usage{
\method{fmtcatch}{nagasaki}(path, spcs, type = NULL)
}
\description{
Make catch data reading multiple Excel worksheets
}
