% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_legacy.R
\name{legacy}
\alias{legacy}
\title{Constructor for legacy Excel data}
\usage{
legacy(fname, year, type, spcs)
}
\arguments{
\item{fname}{Legacy Excel data}

\item{year}{Newest year reported}

\item{spcs}{Spcies name such as: \itemize{\item{'maaji'}}}
}
\description{
Constructor for legacy Excel data
}
