% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_legacy.R
\name{xl2df}
\alias{xl2df}
\title{Convert legacy Excel to data frome}
\usage{
xl2df(fname, year, spcs)
}
\arguments{
\item{fname}{Legacy Excel data}

\item{year}{Newest year reported}

\item{spcs}{Spcies name such as: \itemize{\item{'maaji'}}}
}
\description{
Convert legacy Excel to data frome
}
